@@ -260,7 +260,7 @@
     "ts-loader": "^8.0.11",
     "ts-node": "~9.0.0",
     "tslint": "~6.1.3",
-    "typescript": "~4.0.5",
+    "typescript": "~4.1.0",
     "validate-commit-msg": "^2.14.0",
     "web-animations-js": "^2.3.2",
     "webpack-bundle-analyzer": "^3.0.3",
