@@ -4,8 +4,7 @@ import { environment } from '../../../../environments/environment';
 import { NzModalService } from 'ng-zorro-antd/modal';
 import { TranslateService } from '@ngx-translate/core';
 import { SheetImportPopupComponent } from '../sheet-import-popup/sheet-import-popup.component';
-import { distinctUntilChanged, filter, map, shareReplay, switchMap } from 'rxjs/operators';
-import { LazyDataService } from '../../../core/data/lazy-data.service';
+import { distinctUntilChanged, filter, map, shareReplay } from 'rxjs/operators';
 import { combineLatest } from 'rxjs';
 import { AuthFacade } from '../../../+state/auth.facade';
 import { AllaganReportQueueEntry } from '../model/allagan-report-queue-entry';
@@ -13,6 +12,7 @@ import { NzMessageService } from 'ng-zorro-antd/message';
 import { AllaganReportStatus } from '../model/allagan-report-status';
 import { AllaganReportSource } from '../model/allagan-report-source';
 import { uniq } from 'lodash';
+import { LazyDataFacade } from '../../../lazy-data/+state/lazy-data.facade';
 
 @Component({
   selector: 'app-allagan-reports',
@@ -52,31 +52,32 @@ export class AllaganReportsComponent {
     shareReplay(1)
   );
 
-  public status$ = this.isDataChecker$.pipe(
-    switchMap(dataChecker => {
-      return combineLatest([this.lazyData.extracts$, this.allaganReportsService.getDashboardData()]).pipe(
-        map(([extracts, dashboardData]) => {
-          const fishWithNoData = this.lazyData.data.fishes
-            .filter(itemId => {
-              return !this.lazyData.data.items[itemId].en.includes('Skybuilders')
-                && this.lazyData.getExtract(itemId).sources.length === 0;
-            });
-          return {
-            reportsCount: dashboardData.reportsCount,
-            appliedReportsCount: dashboardData.appliedReportsCount,
-            fishCoverage: Math.floor(1000 * (this.lazyData.data.fishes.length - fishWithNoData.length) / this.lazyData.data.fishes.length) / 10,
-            fishWithNoData,
-            itemsWithNoSource: Object.values(extracts).filter(e => {
-              const enName = this.lazyData.data.items[e.id].en;
-              const frName = this.lazyData.data.items[e.id].fr;
-              return !fishWithNoData.includes(e.id)
-                && !['Dated', 'Skybuilders'].some(ignored => enName.indexOf(ignored) > -1)
-                && !/S\d{1,2}$/.test(frName) && enName.length > 0
-                && e.sources.length === 0;
-            }).map(e => e.id).sort((a, b) => b - a)
-          };
-        })
-      );
+  public status$ = combineLatest([
+    this.lazyData.getEntry('extracts'),
+    this.lazyData.getEntry('fishes'),
+    this.lazyData.getEntry('items'),
+    this.allaganReportsService.getDashboardData()
+  ]).pipe(
+    map(([extracts, fishes, items, dashboardData]) => {
+      const fishWithNoData = fishes
+        .filter(itemId => {
+          return !items[itemId].en.includes('Skybuilders')
+            && extracts[itemId].sources.length === 0;
+        });
+      return {
+        reportsCount: dashboardData.reportsCount,
+        appliedReportsCount: dashboardData.appliedReportsCount,
+        fishCoverage: Math.floor(1000 * (fishes.length - fishWithNoData.length) / fishes.length) / 10,
+        fishWithNoData,
+        itemsWithNoSource: Object.values(extracts).filter(e => {
+          const enName = items[e.id].en;
+          const frName = items[e.id].fr;
+          return !fishWithNoData.includes(e.id)
+            && !['Dated', 'Skybuilders'].some(ignored => enName.indexOf(ignored) > -1)
+            && !/S\d{1,2}$/.test(frName) && enName.length > 0
+            && e.sources.length === 0;
+        }).map(e => e.id).sort((a, b) => b - a)
+      };
     })
   );
 
@@ -85,7 +86,7 @@ export class AllaganReportsComponent {
 
   constructor(public allaganReportsService: AllaganReportsService,
               private dialog: NzModalService, public translate: TranslateService,
-              private lazyData: LazyDataService, private authFacade: AuthFacade,
+              private lazyData: LazyDataFacade, private authFacade: AuthFacade,
               private message: NzMessageService, private cd: ChangeDetectorRef) {
   }
 
