@@ -1,15 +1,13 @@
-import { ChangeDetectionStrategy, Component, TemplateRef, ViewChild } from '@angular/core';
-import { combineLatest, concat, merge, Observable, of, Subject } from 'rxjs';
+import { ChangeDetectionStrategy, Component } from '@angular/core';
+import { combineLatest, merge, Observable, of, Subject } from 'rxjs';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
-import { filter, first, map, mergeMap, tap } from 'rxjs/operators';
+import { filter, first, map, mergeMap, switchMap, tap } from 'rxjs/operators';
 import { AuthFacade } from '../../../+state/auth.facade';
-import { LazyDataService } from '../../../core/data/lazy-data.service';
 import { RotationPickerService } from '../../../modules/rotations/rotation-picker.service';
 import { ListPickerService } from '../../../modules/list-picker/list-picker.service';
 import { ListManagerService } from '../../../modules/list/list-manager.service';
 import { ProgressPopupService } from '../../../modules/progress-popup/progress-popup.service';
 import { ListsFacade } from '../../../modules/list/+state/lists.facade';
-import { List } from '../../../modules/list/model/list';
 import { NzNotificationService } from 'ng-zorro-antd/notification';
 import { I18nToolsService } from '../../../core/tools/i18n-tools.service';
 import { LocalizedDataService } from '../../../core/data/localized-data.service';
@@ -20,6 +18,8 @@ import { AlarmsFacade } from '../../../core/alarms/+state/alarms.facade';
 import { AlarmDisplay } from '../../../core/alarms/alarm-display';
 import { Alarm } from '../../../core/alarms/alarm';
 import { AlarmGroup } from '../../../core/alarms/alarm-group';
+import { LazyDataFacade } from '../../../lazy-data/+state/lazy-data.facade';
+import { safeCombineLatest } from '../../../core/rxjs/safe-combine-latest';
 
 @Component({
   selector: 'app-collectables',
@@ -45,13 +45,6 @@ export class CollectablesComponent {
 
   public levelsEditor$ = new Subject<Record<number, number>>();
 
-  @ViewChild('notificationRef', { static: true })
-  notification: TemplateRef<any>;
-
-  itemsAdded = 0;
-
-  modifiedList: List;
-
   selectedItems: Record<number, number[]> = {};
 
   loading$ = this.authFacade.loaded$.pipe(map(loaded => !loaded));
@@ -59,7 +52,7 @@ export class CollectablesComponent {
   alarmGroups$ = this.alarmsFacade.allGroups$;
 
   selectedTabFromRoute$: Observable<number> = this.activeRoute.paramMap.pipe(
-    map(params => {
+    switchMap(params => {
       const jobAbbr = params.get('jobAbbr') || 'CRP';
       return this.lazyData.getJobIdByAbbr(jobAbbr);
     }),
@@ -77,12 +70,14 @@ export class CollectablesComponent {
   selectedTab$: Observable<number> = merge(this.selectedTabFromRoute$, this.selectedTabFromTabset$);
 
   constructor(private fb: FormBuilder, private authFacade: AuthFacade,
-              private lazyData: LazyDataService, private rotationPicker: RotationPickerService,
+              private lazyData: LazyDataFacade, private rotationPicker: RotationPickerService,
               private listPicker: ListPickerService, private listManager: ListManagerService,
               private progressService: ProgressPopupService, private notificationService: NzNotificationService,
               private listsFacade: ListsFacade, private i18n: I18nToolsService, private l12n: LocalizedDataService,
               private router: Router, private activeRoute: ActivatedRoute, private location: Location,
               private gatheringNodesService: GatheringNodesService, private alarmsFacade: AlarmsFacade) {
+
+    this.lazyData.preloadEntry('paramGrow');
 
     this.form$ = this.levels$.pipe(
       map(levels => {
@@ -99,10 +94,20 @@ export class CollectablesComponent {
         return Object.keys(levels).reduce((res, jobId) => {
           return [
             ...res,
-            { job: jobId, level: +levels[jobId], groups: this.getCollectables(+jobId, +levels[jobId]) }
+            { job: jobId, level: +levels[jobId] }
           ];
         }, []);
-      })
+      }),
+      switchMap(rows => {
+        return safeCombineLatest(rows.map(({ job, level }) => {
+          return this.getCollectables(+job, level).pipe(
+            map(groups => {
+              return { job, level, groups };
+            })
+          );
+        }));
+      }),
+      tap(() => console.log('AYAI'))
     );
   }
 
@@ -144,55 +149,82 @@ export class CollectablesComponent {
     }));
   }
 
-  private getCollectables(jobId: number, level: number): any[] {
-    return Object.keys(this.lazyData.data.collectables)
-      .filter(key => {
-        const collectableEntry = this.lazyData.data.collectables[key];
-        if (collectableEntry.hwd || !collectableEntry.collectable) {
-          return false;
-        }
-        const job = Object.keys(this.lazyData.data.collectablesShops).find(sKey => {
-          return this.lazyData.data.collectablesShops[sKey].indexOf(collectableEntry.shopId) > -1;
-        });
-        return job !== undefined && (+job + 8) === jobId && collectableEntry.levelMin <= level;
-      })
-      .map(key => {
-        return {
-          ...this.lazyData.data.collectables[key],
-          itemId: +key,
-          amount: 1
-        };
-      })
-      .reduce((acc, row) => {
-        let group = acc.find(accRow => accRow.groupId === row.group);
-        if (group === undefined) {
-          acc.push({
-            groupId: row.group,
-            collectables: []
-          });
-          group = acc[acc.length - 1];
-        }
-        group.collectables.push(row);
-        return acc;
-      }, [])
-      .map(group => {
-        group.collectables = group.collectables
-          .sort((a, b) => b.levelMax - a.levelMax)
-          .map(collectable => {
-            if ([16, 17, 18].includes(jobId)) {
-              collectable.nodes = this.gatheringNodesService.getItemNodes(collectable.itemId, true)
-                .map(gatheringNode => {
-                  return {
-                    gatheringNode,
-                    alarms: gatheringNode.limited ? this.alarmsFacade.generateAlarms(gatheringNode) : []
-                  };
-                });
+  private getCollectables(jobId: number, level: number): Observable<any[]> {
+    return combineLatest([
+      this.lazyData.getEntry('collectables'),
+      this.lazyData.getEntry('collectablesShops')
+    ]).pipe(
+      switchMap(([collectables, collectablesShops]) => {
+        return combineLatest(Object.keys(collectables)
+          .filter(key => {
+            const collectableEntry = collectables[key];
+            if (collectableEntry.hwd || !collectableEntry.collectable) {
+              return false;
             }
-            return collectable;
-          });
-        return group;
+            const job = Object.keys(collectablesShops).find(sKey => {
+              return collectablesShops[sKey].includes(collectableEntry.shopId);
+            });
+            return job !== undefined && (+job + 8) === jobId && collectableEntry.levelMin <= level;
+          })
+          .map(key => {
+            return {
+              ...collectables[key],
+              itemId: +key,
+              amount: 1
+            };
+          })
+          .reduce((acc, row) => {
+            let group = acc.find(accRow => accRow.groupId === row.group);
+            if (group === undefined) {
+              acc.push({
+                groupId: row.group,
+                collectables: []
+              });
+              group = acc[acc.length - 1];
+            }
+            group.collectables.push(row);
+            return acc;
+          }, [])
+          .map(group => {
+            return safeCombineLatest(group.collectables
+              .sort((a, b) => b.levelMax - a.levelMax)
+              .map(collectable => {
+                return combineLatest([
+                  this.getExp(level, collectable, collectable.base.exp),
+                  this.getExp(level, collectable, collectable.mid.exp),
+                  this.getExp(level, collectable, collectable.high.exp)
+                ]).pipe(
+                  switchMap(([expBase, expMid, expHigh]) => {
+                    collectable.expBase = expBase;
+                    collectable.expMid = expMid;
+                    collectable.expHigh = expHigh;
+                    if ([16, 17, 18].includes(jobId)) {
+                      return this.gatheringNodesService.getItemNodes(collectable.itemId, true).pipe(
+                        map(nodes => {
+                          collectable.nodes = nodes.map(gatheringNode => {
+                            return {
+                              gatheringNode,
+                              alarms: gatheringNode.limited ? this.alarmsFacade.generateAlarms(gatheringNode) : []
+                            };
+                          });
+                          return collectable;
+                        })
+                      );
+                    }
+                    return of(collectable);
+                  })
+                );
+              })
+            ).pipe(
+              map(res => {
+                group.collectables = res;
+                return group;
+              })
+            );
+          })
+        );
       })
-      .reverse();
+    );
   }
 
   toggleAlarm(display: AlarmDisplay): void {
@@ -208,80 +240,58 @@ export class CollectablesComponent {
   }
 
   public openInSimulator(itemId: number): void {
-    this.rotationPicker.openInSimulator(itemId, this.lazyData.getItemRecipeSync(itemId.toString())?.id);
+    this.lazyData.getRecipeForItem(itemId).pipe(
+      first()
+    ).subscribe(recipe => {
+      this.rotationPicker.openInSimulator(itemId, recipe?.id);
+    });
   }
 
   public addItemsToList(items: { itemId: number, amount: number, collectable?: boolean }[]): void {
-    this.listPicker.pickList().pipe(
-      mergeMap(list => {
-        const operations = items.map(item => {
-          const recipeId = this.lazyData.getItemRecipeSync(item.itemId.toString())?.id;
-          return this.listManager.addToList({
-            itemId: +item.itemId,
-            list: list,
-            recipeId: recipeId || '',
+    combineLatest(items.map(item => {
+      return this.lazyData.getRecipeForItem(item.itemId).pipe(
+        map(recipe => {
+          return {
+            id: +item.itemId,
+            recipeId: recipe?.id || '',
             amount: item.amount,
             collectable: item.collectable
-          });
-        });
-        let operation$: Observable<any>;
-        if (operations.length > 0) {
-          operation$ = concat(
-            ...operations
-          );
-        } else {
-          operation$ = of(list);
-        }
-        return this.progressService.showProgress(operation$,
-          items.length,
-          'Adding_recipes',
-          { amount: items.length, listname: list.name });
-      }),
-      tap(list => list.$key ? this.listsFacade.updateList(list) : this.listsFacade.addList(list)),
-      mergeMap(list => {
-        // We want to get the list created before calling it a success, let's be pessimistic !
-        return this.progressService.showProgress(
-          combineLatest([this.listsFacade.myLists$, this.listsFacade.listsWithWriteAccess$]).pipe(
-            map(([myLists, listsICanWrite]) => [...myLists, ...listsICanWrite]),
-            map(lists => lists.find(l => l.createdAt.toMillis() === list.createdAt.toMillis())),
-            filter(l => l !== undefined),
-            first()
-          ), 1, 'Saving_in_database');
-      })
-    ).subscribe((list) => {
-      this.itemsAdded = items.length;
-      this.modifiedList = list;
-      this.notificationService.template(this.notification);
-    });
+          };
+        })
+      );
+    })).pipe(
+      switchMap(additions => this.listPicker.addToList(...additions))
+    );
   }
 
   public createQuickList(item: { itemId: number, amount: number }): void {
     const list = this.listsFacade.newEphemeralList(this.i18n.getName(this.l12n.getItem(+item.itemId)));
-    const recipeId = this.lazyData.getItemRecipeSync(item.itemId.toString())?.id;
-    const operation$ = this.listManager.addToList({
-      itemId: +item.itemId,
-      list: list,
-      recipeId: recipeId || '',
-      amount: item.amount
-    })
-      .pipe(
-        tap(resultList => this.listsFacade.addList(resultList)),
-        mergeMap(resultList => {
-          return this.listsFacade.myLists$.pipe(
-            map(lists => lists.find(l => l.createdAt.toMillis() === resultList.createdAt.toMillis() && l.$key !== undefined)),
-            filter(l => l !== undefined),
-            first()
-          );
-        })
-      );
+    this.lazyData.getRecipeForItem(item.itemId).pipe(
+      switchMap(recipe => {
+        const operation$ = this.listManager.addToList({
+          itemId: +item.itemId,
+          list: list,
+          recipeId: recipe?.id || '',
+          amount: item.amount
+        }).pipe(
+          tap(resultList => this.listsFacade.addList(resultList)),
+          mergeMap(resultList => {
+            return this.listsFacade.myLists$.pipe(
+              map(lists => lists.find(l => l.createdAt.toMillis() === resultList.createdAt.toMillis() && l.$key !== undefined)),
+              filter(l => l !== undefined),
+              first()
+            );
+          })
+        );
 
-    this.progressService.showProgress(operation$, 1)
-      .subscribe((newList) => {
-        this.router.navigate(['list', newList.$key]);
-      });
+        return this.progressService.showProgress(operation$, 1);
+      })
+    ).subscribe((newList) => {
+      this.router.navigate(['list', newList.$key]);
+    });
   }
 
-  public getExp(level: number, collectable: any, ratio: number): number {
+  public getExp(level: number, collectable: any, ratio: number): Observable<number> {
     const firstCollectableDigit = Math.floor(collectable.levelMax / 10);
     const firstLevelDigit = Math.floor(level / 10);
     let nerfedExp = firstCollectableDigit < firstLevelDigit;
@@ -291,9 +301,11 @@ export class CollectablesComponent {
         || collectable.levelMax % 10 === 0;
     }
     if (nerfedExp) {
-      return 10000;
+      return of(10000);
     }
-    return this.lazyData.data.paramGrow[collectable.levelMax].ExpToNext * ratio / 1000;
+    return this.lazyData.getRow('paramGrow', collectable.levelMax).pipe(
+      map(row => row.ExpToNext * ratio / 1000)
+    );
   }
 
 }
