@@ -1,7 +1,7 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { LazyDataService } from 'apps/client/src/app/core/data/lazy-data.service';
-import { combineLatest, Observable } from 'rxjs';
-import { filter, map, shareReplay } from 'rxjs/operators';
+import { combineLatest, Observable, of } from 'rxjs';
+import { filter, map, shareReplay, switchMap } from 'rxjs/operators';
 import { FishContextService } from '../../service/fish-context.service';
 import { GatheringNodesService } from '../../../../core/data/gathering-nodes.service';
 import { AlarmsFacade } from 'apps/client/src/app/core/alarms/+state/alarms.facade';
@@ -19,17 +19,23 @@ import { AuthFacade } from 'apps/client/src/app/+state/auth.facade';
 export class FishingSpotAvailableFishesComponent {
   public readonly fishes$: Observable<{ itemId: number, alarms: Alarm[], done: boolean }[] | undefined> = combineLatest([this.fishCtx.spotId$, this.lazyData.fishingSpots$, this.authFacade.logTracking$]).pipe(
     filter(([spotId]) => spotId >= 0),
-    map(([spotId, spots, logs]) => {
+    switchMap(([spotId, spots, logs]) => {
       const fishIds = spots.find((s) => s.id === spotId)?.fishes?.filter((f) => f > 0);
-      return fishIds.map(itemId => {
-        const nodes = this.gatheringNodesService.getItemNodes(itemId, true);
-        const spotNode = nodes.find(n => n.id === spotId);
-        return {
-          itemId,
-          alarms: this.alarmsFacade.generateAlarms(spotNode),
-          done: logs.gathering.includes(itemId)
-        };
-      });
+      if (fishIds.length === 0) {
+        return of([]);
+      }
+      return combineLatest(fishIds.map(itemId => {
+        return this.gatheringNodesService.getItemNodes(itemId, true).pipe(
+          map(nodes => {
+            const spotNode = nodes.find(n => n.id === spotId);
+            return {
+              itemId,
+              alarms: this.alarmsFacade.generateAlarms(spotNode),
+              done: logs.gathering.includes(itemId)
+            };
+          })
+        );
+      }));
     }),
     shareReplay(1)
   );
