@@ -4,24 +4,24 @@ export interface GatheringNode {
   id: number;
   items: number[];
   hiddenItems?: number[];
-  limited: boolean;
+  limited?: boolean;
   level: number;
   type: number;
-  legendary: boolean;
-  ephemeral: boolean;
+  legendary?: boolean;
+  ephemeral?: boolean;
   spawns?: number[];
   duration?: number;
   zoneId: number;
-  map: number;
-  x: number;
-  y: number;
-  z: number;
+  map?: number;
+  x?: number;
+  y?: number;
+  z?: number;
   folklore?: number;
   /**
    * Fishing stuff
    */
   predators?: { id: number, amount: number }[];
-  hookset: number;
+  hookset?: number;
   baits?: FishingBait[];
   weathers?: number[];
   weathersFrom?: number[];
