@@ -1,6 +1,6 @@
 export interface LazyPlace {
-  en: number | string;
-  ja: number | string;
-  de: number | string;
-  fr: number | string;
+  en: string;
+  ja: string;
+  de: string;
+  fr: string;
 }
