@@ -1,91 +1,110 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { Component, Input } from '@angular/core';
 import { TeamcraftGearset } from '../../../model/gearset/teamcraft-gearset';
-import { LazyDataService } from '../../../core/data/lazy-data.service';
-import { getItemSource } from '../../list/model/list-row';
+import { getItemSource, ListRow } from '../../list/model/list-row';
 import { DataType } from '../../list/data/data-type';
 import { TradeSource } from '../../list/model/trade-source';
 import { TranslateService } from '@ngx-translate/core';
 import { TradeEntry } from '../../list/model/trade-entry';
 import { GearsetProgression } from '../../../model/gearset/gearset-progression';
+import { safeCombineLatest } from '../../../core/rxjs/safe-combine-latest';
+import { LazyDataFacade } from '../../../lazy-data/+state/lazy-data.facade';
+import { map, switchMap, tap } from 'rxjs/operators';
+import { combineLatest, Observable } from 'rxjs';
+import { LazyData } from '../../../lazy-data/lazy-data';
+import { observeInput } from '../../../core/rxjs/observe-input';
 
 @Component({
   selector: 'app-gearset-cost-popup',
   templateUrl: './gearset-cost-popup.component.html',
   styleUrls: ['./gearset-cost-popup.component.less']
 })
-export class GearsetCostPopupComponent implements OnInit {
+export class GearsetCostPopupComponent {
 
   @Input()
   gearset: TeamcraftGearset;
 
-  private _progression: GearsetProgression;
-
   @Input()
-  set progression(p: GearsetProgression) {
-    this._progression = p;
-    this.computeCosts();
+  progression: GearsetProgression;
+
+  costs$: Observable<{ id: string | number, amount: number }[]> = combineLatest([
+    observeInput(this)('gearset'),
+    observeInput(this)('progression')
+  ]).pipe(
+    switchMap(([gearset, progression]) => {
+      return safeCombineLatest(Object.keys(gearset)
+        .filter(key => {
+          const validPiece = gearset[key] && gearset[key].itemId !== undefined;
+          const notDone = !progression || !progression[key]?.item;
+          return validPiece && notDone;
+        })
+        .map(key => {
+          const gearPiece = gearset[key];
+          return this.lazyData.getRow('extracts', gearPiece.itemId).pipe(
+            map(itemExtract => {
+              const trades = getItemSource<TradeSource[]>(itemExtract, DataType.TRADE_SOURCES);
+              return trades.filter(trade => trade.npcs.length > 0)
+                .map(trade => {
+                  return trade.trades.map(t => {
+                    return t.currencies;
+                  }).flat();
+                }).flat();
+            })
+          );
+        })
+      );
+    }),
+    switchMap((currencies) => {
+      return combineLatest([
+        this.lazyData.getEntry('extracts'),
+        this.lazyData.getEntry('itemEquipSlotCategory')
+      ]).pipe(
+        map(([extracts, itemEquipSlotCategory]) => {
+          return this.mergeCosts(...currencies.flat().map(currency => {
+            return this.computeCosts(currency, itemEquipSlotCategory, extracts);
+          }).flat());
+        })
+      );
+    })
+  );
+
+  constructor(private lazyData: LazyDataFacade, public translate: TranslateService) {
   }
 
-  get progression(): GearsetProgression {
-    return this._progression;
-  }
-
-  costs: { id: string | number, amount: number }[] = [];
-
-  constructor(private lazyData: LazyDataService, public translate: TranslateService) {
-  }
-
-  ngOnInit(): void {
-    this.computeCosts();
-  }
-
-  private computeCosts(): void {
-    this.costs = [];
-    Object.keys(this.gearset)
-      .filter(key => {
-        const validPiece = this.gearset[key] && this.gearset[key].itemId !== undefined;
-        const notDone = !this.progression || !this.progression[key]?.item;
-        return validPiece && notDone;
-      })
-      .forEach(key => {
-        const gearPiece = this.gearset[key];
-        const itemExtract = this.lazyData.getExtract(gearPiece.itemId);
-        const trades = getItemSource<TradeSource[]>(itemExtract, DataType.TRADE_SOURCES);
-        trades.filter(trade => trade.npcs.length > 0).forEach(trade => {
-          trade.trades.forEach(t => {
-            t.currencies.forEach(currency => {
-              this.addCurrency(currency);
-            });
-          });
-        });
-      });
-  }
-
-  private addCurrency(currency: TradeEntry): void {
+  private computeCosts(currency: TradeEntry, itemEquipSlotCategory: LazyData['itemEquipSlotCategory'], extracts: Record<number, ListRow>, costs = []): { id: string | number, amount: number }[] {
     // If you can equip this item, let's trigger recursion and not include it as a trade currency
-    if (this.lazyData.data.itemEquipSlotCategory[currency.id]) {
-      const itemExtract = this.lazyData.getExtract(+currency.id);
+    if (itemEquipSlotCategory[currency.id]) {
+      const itemExtract = extracts[+currency.id];
       const trades = getItemSource<TradeSource[]>(itemExtract, DataType.TRADE_SOURCES);
       if (trades.length > 0) {
-        trades.filter(trade => trade.npcs.length > 0).forEach(trade => {
-          trade.trades.forEach(t => {
-            t.currencies.forEach(c => {
-              this.addCurrency(c);
-            });
-          });
-        });
-        return;
+        const tradeCurrencyCosts = trades
+          .filter(trade => trade.npcs.length > 0)
+          .map(trade => {
+            return trade.trades.map(t => {
+              return t.currencies.map(c => {
+                return this.computeCosts(c, itemEquipSlotCategory, extracts);
+              }).flat();
+            }).flat();
+          }).flat();
+        return this.mergeCosts(...costs, ...tradeCurrencyCosts);
       }
     }
-    let costRow = this.costs.find(c => c.id === currency.id);
-    if (costRow === undefined) {
-      this.costs.push({
-        id: currency.id,
-        amount: 0
-      });
-      costRow = this.costs[this.costs.length - 1];
-    }
-    costRow.amount += currency.amount;
+    return this.mergeCosts(...costs, { id: currency.id, amount: currency.amount });
+  }
+
+  private mergeCosts(...costs: { id: string | number, amount: number }[]): { id: string | number, amount: number }[] {
+    const newCosts = [];
+    costs.forEach(cost => {
+      let costRow = newCosts.find(c => c.id === cost.id);
+      if (costRow === undefined) {
+        newCosts.push({
+          id: cost.id,
+          amount: 0
+        });
+        costRow = newCosts[newCosts.length - 1];
+      }
+      costRow.amount += cost.amount;
+    });
+    return newCosts;
   }
 
 }
