@@ -50,7 +50,7 @@ export class ListPickerService {
       );
   }
 
-  addToList(...items: ListRow[]): Observable<List> {
+  addToList(...items: Pick<ListRow, 'id' | 'recipeId' | 'amount'>[]): Observable<List> {
     // Making the observable optional, this way you can just call it and ignore it,
     // or add your own logic once it's done.
     const done$ = new Subject<List>();
@@ -89,6 +89,8 @@ export class ListPickerService {
           ), 1, 'Saving_in_database');
       }),
       switchMap(list => {
+        done$.next(list);
+        done$.complete();
         return this.notificationService.success(
           this.translate.instant('Success'),
           this.translate.instant('Recipes_Added', { listname: list.name, itemcount: items.length }),
@@ -103,8 +105,6 @@ export class ListPickerService {
       })
     ).subscribe((list) => {
       this.router.navigate(['/list', list.$key]);
-      done$.next(list);
-      done$.complete();
     });
     return done$;
   }
