@@ -7,6 +7,8 @@ import { ListRow } from '../../model/list-row';
 import { GarlandToolsService } from '../../../../core/api/garland-tools.service';
 import { AlarmsFacade } from '../../../../core/alarms/+state/alarms.facade';
 import { GatheringNodesService } from '../../../../core/data/gathering-nodes.service';
+import { map } from 'rxjs/operators';
+import { Observable } from 'rxjs';
 
 export class AlarmsExtractor extends AbstractExtractor<Alarm[]> {
   constructor(gt: GarlandToolsService, private gatheringNodesService: GatheringNodesService,
@@ -19,16 +21,19 @@ export class AlarmsExtractor extends AbstractExtractor<Alarm[]> {
   }
 
   isAsync(): boolean {
-    return false;
+    return true;
   }
 
   protected canExtract(item: Item): boolean {
     return true;
   }
 
-  protected doExtract(item: Item, itemData: ItemData, row: ListRow): Alarm[] {
-    const nodes = this.gatheringNodesService.getItemNodes(item.id);
-    return nodes.map(node => node.limited ? this.alarmsFacade.generateAlarms(node) : []).flat();
+  protected doExtract(item: Item, itemData: ItemData, row: ListRow): Observable<Alarm[]> {
+    return this.gatheringNodesService.getItemNodes(item.id).pipe(
+      map(nodes => {
+        return nodes.map(node => node.limited ? this.alarmsFacade.generateAlarms(node) : []).flat();
+      })
+    );
   }
 
 }
