@@ -76,57 +76,57 @@ const extractors: AbstractExtractor[] = [
   new I18nExtractor('SpecialShop', 'shops'),
   new I18nExtractor('AirshipExplorationPoint', 'airship-voyages', { ID: 'id' }, 'NameShort_', true),
   new I18nExtractor('SubmarineExploration', 'submarine-voyages', { ID: 'id' }, 'Destination_'),
-  new SeedsExtractor(),
-  new WorldsExtractor(),
-  new TerritoriesExtractor(),
-  new CollectablesExtractor(),
-  new HwdGathererExtractor(),
-  new ActionTimelineExtractor(),
-  new MateriasExtractor(),
-  new BaseParamExtractor(),
-  new ItemLevelExtractor(),
-  new ClassJobModifiersExtractor(),
-  new EquipSlotCategoryExtractor(),
-  new TribesExtractor(),
-  new VenturesExtractor(),
-  new FoodsExtractor(),
-  new MedicinesExtractor(),
-  new ParamGrowExtractor(),
-  new CollectablesShopsExtractor(),
-  new NpcsExtractor(),
-  new LevesExtractor(),
-  new JobsExtractor(),
-  new JobCategoriesExtractor(),
-  new FatesExtractor(),
-  new GatheringBonusesExtractor(),
-  new CdGroupsExtractor(),
-  new CombosExtractor(),
-  new ItemsExtractor(),
-  new AetherytesExtractor(),
-  new RecipesExtractor(),
-  new ActionsExtractor(),
-  new ReductionsExtractor(),
-  new MonsterDropsExtractor(),
-  new StatsExtractor(),
-  new InstancesExtractor(),
-  new QuestsExtractor(),
-  new MapsExtractor(),
-  new MapIdsExtractor(),
-  new FishParameterExtractor(),
-  new AetherstreamExtractor(),
-  new WeathersExtractor(),
-  new WeatherRateExtractor(),
-  new PatchContentExtractor(),
-  new LogsExtractor(),
-  new SubmarinePartsExtractor(),
-  new SubmarineRanksExtractor(),
-  new AirshipPartsExtractor(),
-  new AirshipRanksExtractor(),
-  new TreasuresExtractor(),
-  new MappyExtractor(),
-  new LgbExtractor(),
-  new GubalExtractor(),
-  new AllaganReportsExtractor()
+  // new SeedsExtractor(),
+  // new WorldsExtractor(),
+  // new TerritoriesExtractor(),
+  // new CollectablesExtractor(),
+  // new HwdGathererExtractor(),
+  // new ActionTimelineExtractor(),
+  // new MateriasExtractor(),
+  // new BaseParamExtractor(),
+  // new ItemLevelExtractor(),
+  // new ClassJobModifiersExtractor(),
+  // new EquipSlotCategoryExtractor(),
+  // new TribesExtractor(),
+  // new VenturesExtractor(),
+  // new FoodsExtractor(),
+  // new MedicinesExtractor(),
+  // new ParamGrowExtractor(),
+  // new CollectablesShopsExtractor(),
+  // new NpcsExtractor(),
+  // new LevesExtractor(),
+  // new JobsExtractor(),
+  // new JobCategoriesExtractor(),
+  // new FatesExtractor(),
+  // new GatheringBonusesExtractor(),
+  // new CdGroupsExtractor(),
+  // new CombosExtractor(),
+  // new ItemsExtractor(),
+  // new AetherytesExtractor(),
+  // new RecipesExtractor(),
+  // new ActionsExtractor(),
+  // new ReductionsExtractor(),
+  // new MonsterDropsExtractor(),
+  // new StatsExtractor(),
+  // new InstancesExtractor(),
+  // new QuestsExtractor(),
+  // new MapsExtractor(),
+  // new MapIdsExtractor(),
+  // new FishParameterExtractor(),
+  // new AetherstreamExtractor(),
+  // new WeathersExtractor(),
+  // new WeatherRateExtractor(),
+  // new PatchContentExtractor(),
+  // new LogsExtractor(),
+  // new SubmarinePartsExtractor(),
+  // new SubmarineRanksExtractor(),
+  // new AirshipPartsExtractor(),
+  // new AirshipRanksExtractor(),
+  // new TreasuresExtractor(),
+  // new MappyExtractor(),
+  // new LgbExtractor(),
+  // new GubalExtractor(),
+  // new AllaganReportsExtractor()
 ];
 
 if (process.env.XIVAPI_KEY) {
